requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pypdfium2>=4.0.0
Pillow>=10.0.0
streamlit>=1.30.0
requests>=2.31.0
